Finally renamed binaries
